+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -1,0 +1,12 @@
+//
+//  Generated file. Do not edit.
+//
+
+import FlutterMacOS
+import Foundation
+
+import liquid_glass_container
+
+func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  LiquidGlassContainerPlugin.register(with: registry.registrar(forPlugin: "LiquidGlassContainerPlugin"))
+}